test: create email validation and add to composite
