+++ src/main/factories/signup-validation.spec.ts
@@ -1,10 +1,23 @@
 import { CompareFieldsValidation } from "../../presentation/helpers/validators/compare-fields-validation"
+import { EmailValidation } from "../../presentation/helpers/validators/email-validation"
 import { RequiredFieldValidation } from "../../presentation/helpers/validators/required-field-validation"
 import { Validation } from "../../presentation/helpers/validators/validation"
 import { ValidationComposite } from "../../presentation/helpers/validators/validation-composite"
+import { EmailValidator } from "../../presentation/protocols/email-validator"
 import { makeSignUpValidation } from "./signup-validation"
 
 jest.mock("../../presentation/helpers/validators/validation-composite")
+
+const makeEmailValidator = (): EmailValidator => {
+  class EmailValidatorStub implements EmailValidator {
+    isValid(email: string): boolean {
+      console.log(email)
+      return true
+    }
+  }
+
+  return new EmailValidatorStub()
+}
 
 describe("SignUpValidation Factory", () => {
   test("Should call validation composite with all validators", () => {
@@ -18,6 +31,8 @@ describe("SignUpValidation Factory", () => {
     validations.push(
       new CompareFieldsValidation("password", "passwordConfirmation")
     )
+
+    validations.push(new EmailValidation("email", makeEmailValidator()))
 
     expect(ValidationComposite).toHaveBeenCalledWith(validations)
   })
