+++ src/presentation/controllers/signup/signup.spec.ts
@@ -1,31 +1,18 @@
-import { InvalidParamError, MissingParamError, ServerError } from "../../errors"
+import { MissingParamError, ServerError } from "../../errors"
 import { badRequest, ok, serverError } from "../../helpers/http-helper"
 import { SignUpController } from "./signup"
 import {
   AccountModel,
   AddAccount,
   AddAccountModel,
-  EmailValidator,
   HttpRequest,
   Validation,
 } from "./signup-protocols"
 
 interface SutTypes {
   sut: SignUpController
-  emailValidatorStub: EmailValidator
   addAccountStub: AddAccount
   validationStub: Validation
-}
-
-const makeEmailValidator = (): EmailValidator => {
-  class EmailValidatorStub implements EmailValidator {
-    isValid(email: string): boolean {
-      console.log(email)
-      return true
-    }
-  }
-
-  return new EmailValidatorStub()
 }
 
 const makeFakeRequest = (): HttpRequest => ({
@@ -67,53 +54,18 @@ const makeAddAccount = (): AddAccount => {
 }
 
 const makeSut = (): SutTypes => {
-  const emailValidatorStub = makeEmailValidator()
   const addAccountStub = makeAddAccount()
   const validationStub = makeValidation()
-  const sut = new SignUpController(
-    emailValidatorStub,
-    addAccountStub,
-    validationStub
-  )
+  const sut = new SignUpController(addAccountStub, validationStub)
 
   return {
     sut,
-    emailValidatorStub,
     addAccountStub,
     validationStub,
   }
 }
 
 describe("SignUp Controller", () => {
-  test("Should return 400 if an invalid e-mail is provided", async () => {
-    const { sut, emailValidatorStub } = makeSut()
-
-    jest.spyOn(emailValidatorStub, "isValid").mockReturnValueOnce(false)
-
-    const httpResponse = await sut.handle(makeFakeRequest())
-    expect(httpResponse).toEqual(badRequest(new InvalidParamError("email")))
-  })
-
-  test("Should call EmailValidator with the correct email", async () => {
-    const { sut, emailValidatorStub } = makeSut()
-
-    const isValidSpy = jest.spyOn(emailValidatorStub, "isValid")
-
-    await sut.handle(makeFakeRequest())
-    expect(isValidSpy).toHaveBeenCalledWith("any_email@mail.com")
-  })
-
-  test("Should return 500 if EmailValidation throws", async () => {
-    const { sut, emailValidatorStub } = makeSut()
-
-    jest.spyOn(emailValidatorStub, "isValid").mockImplementationOnce(() => {
-      throw new Error()
-    })
-
-    const httpResponse = await sut.handle(makeFakeRequest())
-    expect(httpResponse).toEqual(serverError(new ServerError()))
-  })
-
   test("Should call AddAccount with the correct values", async () => {
     const { sut, addAccountStub } = makeSut()
